requirements.txt
alembic==0.6.5
blinker==1.3
decorator==3.4.0
-e git+https://github.com/pythonforfacebook/facebook-sdk.git@8c0d34291aaafec00e02eaa71cc2a242790a0fcc#egg=facebook_sdk-master
Flask==0.10.1
Flask-Assets==0.10
Flask-DebugToolbar==0.10.0
Flask-Login==0.2.10
Flask-Mail==0.9.1
Flask-Migrate==1.2.0
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.2
Flask-SQLAlchemy==2.0
Flask-WTF==0.9.5
hashids==1.0.3
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
passlib==1.6.2
pbr==0.8.0
py==1.4.30
PyYAML==3.11
requests==2.7.0
six==1.6.1
SQLAlchemy==1.0.6
sqlalchemy-migrate==0.9.1
Tempita==0.5.2
-e git+https://github.com/unifispot/unifi-api/@20d2f8d07244d53523292d6a9eec487c72ec76af#egg=unifi-master
webassets==0.10.1
Werkzeug==0.10.4
wheel==0.24.0
WTForms==1.0.5
